requirements.txt
Django==1.9.8
MySQL-python==1.2.5
argparse==1.2.1
django-extensions==1.6.7
django-registration==2.1.2
six==1.10.0
wsgiref==0.1.2
